refactor(project): extract PiquleDirectory interface
