+++ bin/piqule.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 use Haspadar\Piqule\Output\Console;
 use Haspadar\Piqule\Output\Line\Error;
 use Haspadar\Piqule\PiquleException;
+use Haspadar\Piqule\Project\DiskPiquleDirectory;
 use Haspadar\Piqule\Project\InitializedProject;
 use Haspadar\Piqule\Project\Lock\JsonLock;
-use Haspadar\Piqule\Project\PiquleDirectory;
 use Haspadar\Piqule\Project\ProjectOf;
 use Haspadar\Piqule\Project\UninitializedProject;
 use Haspadar\Piqule\RunContext;
@@ -23,10 +23,8 @@ $output = new Console();
 try {
     $context = new RunContext($argv);
     $root = $context->root();
-    $piqule = new PiquleDirectory($root . '/.piqule');
-    $sourceDirectory = new DiskSourceDirectory(
-        dirname(__DIR__) . '/templates',
-    );
+    $piqule = new DiskPiquleDirectory($root . '/.piqule');
+    $sourceDirectory = new DiskSourceDirectory(dirname(__DIR__) . '/templates');
     $targetDirectory = new DiskTargetDirectory($root);
     $project = new ProjectOf(
         $piqule,
